renaming add-thread to add-channel
