+++ src/app/nip76/nip76-add-channel-dialog/add-channel-dialog.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AddChannelDialog } from './add-thread-dialog.component';
+import { AddChannelDialog } from './add-channel-dialog.component';
 
 describe('AddChannelDialogComponent', () => {
   let component: AddChannelDialog;
